pyproject.toml
[project]
name = "uv-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["mypy[mypyc]>=1.11.2"]

[build-system]
requires = ["setuptools", "wheel", "mypy", "mypyc"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["uv-demo"]
